test(ts): unify `Join Group` benchmark
